Use MultiWriterIDGenerator for device inbox
